# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "XSocial Laravel API",
        "description": "✊✊✊",
        "contact": {
            "email": "ngoctoan06011998@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "comments.index",
                "operationId": "0406f270b674fae17bf23b4206a3004f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "filter[commentable_type]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "POST",
                                    "REPLY_COMMENT"
                                ]
                            }
                        }
                    },
                    {
                        "name": "filter[commentable_id]",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "user|favorites",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "-id, -created_at",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "comments.store",
                "description": "commentable_type: POST, REPLY_COMMENT",
                "operationId": "300619a0a4003920aafe960e45688102",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "commentable_id",
                                    "commentable_type",
                                    "body",
                                    "media_ids"
                                ],
                                "properties": {
                                    "commentable_id": {
                                        "type": "integer"
                                    },
                                    "commentable_type": {
                                        "type": "string"
                                    },
                                    "body": {
                                        "type": "string"
                                    },
                                    "media_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments/{id}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "comments.show",
                "operationId": "ca005bcffa22ab29b155dd7aa13b5d6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "user|favorites",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "comments.update",
                "operationId": "39e62963cbf897a1b2168683878e0a46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "body": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "comments.destroy",
                "operationId": "721e5dc0c32ef81276203df965767606",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "favorites.index",
                "operationId": "81123175f76259d611e2c337477dfa5f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "user",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "-id, -created_at",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "favorites.store",
                "description": "favoritetable_type: POST, COMMENT, MESSAGE",
                "operationId": "370ffab0458b21c048b584dc20d059e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "favoritetable_id",
                                    "favoritetable_type"
                                ],
                                "properties": {
                                    "favoritetable_id": {
                                        "type": "integer"
                                    },
                                    "favoritetable_type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "favorites.destroy",
                "description": "favoritetable_type: POST, COMMENT, MESSAGE",
                "operationId": "fa0c59fb9c161b977258ec7656104ab2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "favoritetable_id",
                                    "favoritetable_type"
                                ],
                                "properties": {
                                    "favoritetable_id": {
                                        "type": "integer"
                                    },
                                    "favoritetable_type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Bearer {token}"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}